Add DM sessions

+++ cogs/linker.py
@@ -9,12 +9,20 @@ class Session:
         self.webhook = webhook
         self.messages = {}
 
+class DMSession:
+    """Represents a sesion bettwen the bot owner and a bot user"""
+    
+    def __init__(self, channel):
+        self.channel = channel
+        self.messages = {}
+
 class Linker(commands.Cog):
     """Linking two channels together"""
 
     def __init__(self, bot):
         self.bot = bot
         self.linked = {}
+        self.dm_sessions = {}
     
     @commands.has_permissions(manage_channels=True)
     @commands.group(name="link", description="Links a channel to another channel", usage="[channel id]", invoke_without_command=True)
@@ -60,30 +68,98 @@ class Linker(commands.Cog):
         self.linked.pop(session.channel)
         await self.bot.get_channel(session.channel).send("✅ Unlinked from the channel")
 
+    @commands.group(name="dm", description="Creates a DM session with a user", usage="[user]", invoke_without_command=True)
+    @commands.guild_only()
+    @commands.is_owner()
+    async def dm_command(self, ctx, *, user: discord.Member):
+        if not user.dm_channel:
+            await user.create_dm()
+
+        if user.dm_channel.id in self.dm_sessions:
+            return await ctx.send("❌ A DM session is already going with this user")
+
+        if ctx.channel.id in self.dm_sessions:
+           return await ctx.send("❌ A DM session is already going in this channel")
+
+        overwrites = {
+            ctx.guild.default_role: discord.PermissionOverwrite(read_messages=False),
+            ctx.guild.me: discord.PermissionOverwrite(read_messages=True, send_messages=True)
+        }
+        channel = await ctx.guild.create_text_channel(name=f"dm-{user}", overwrites=overwrites)
+
+        self.linked[channel.id] = DMSession(user.dm_channel.id)
+        self.linked[user.dm_channel.id] = DMSession(channel.id)
+
+        await ctx.send("✅ Created a DM session")
+
+    @dm_command.command(name="stop", description="Stops a DM session with a user", usage="[user]", aliases=["close"])
+    async def dm_stop_command(self, ctx, *, user: discord.Member):
+        if user.dm_channel.id not in self.linked:
+            return await ctx.send("❌ No DM session with user")
+        
+        session = self.linked.get(user.dm_channel.id)
+        channel = self.bot.get_channel(session.channel)
+        try:
+            await channel.delete()
+        except:
+            pass
+
+        self.linked.pop(user.dm_channel.id)
+        self.linked.pop(ctx.channel.id)
+
     @commands.Cog.listener("on_message")
     async def on_message(self, message):
         session = self.linked.get(message.channel.id)
 
-        if not session or self.linked[session.channel].webhook.id == message.author.id:
+        if not session:
             return
         
-        msg = await session.webhook.send(message.content, username=message.author.display_name, avatar_url=message.author.avatar_url, embeds=message.embeds,wait=True)
-        session.messages[message.id] = msg.id
+        if isinstance(session, Session):
+            if self.linked[session.channel].webhook.id == message.author.id:
+                return
+
+            msg = await session.webhook.send(message.content, username=message.author.display_name, avatar_url=message.author.avatar_url, embeds=message.embeds,wait=True)
+            session.messages[message.id] = msg.id
+        
+        if isinstance(session, DMSession):
+            #Ignore the message if it's the bot
+            #This way, the bot won't spam the same message over and over
+            if message.author.id == self.bot.user.id:
+                return
+
+            destination = self.bot.get_channel(session.channel)
+
+            msg = await destination.send(message.content)
+            session.messages[message.id] = msg.id
 
     @commands.Cog.listener("on_message_delete")
     async def on_message_delete(self, message):
-
         session = self.linked.get(message.channel.id)
 
-        if not session or self.linked[session.channel].webhook.id == message.author.id:
+        if not session:
             return
-
-        destination = self.bot.get_channel(session.channel)
         
-        #Get the deleted message's message in the linked channel and delete it
-        to_delete = await destination.fetch_message(session.messages[message.id])
 
-        await to_delete.delete()
+        if isinstance(session, Session):
+            if self.linked[session.channel].webhook.id == message.author.id:
+                return
+
+            destination = self.bot.get_channel(session.channel)
+            
+            #Get the deleted message's message in the linked channel and delete it
+            to_delete = await destination.fetch_message(session.messages[message.id])
+
+            await to_delete.delete()
+        
+        if isinstance(session, DMSession):
+            #This ignores the bot editing a message
+            if message.author.id == self.bot.user.id:
+                return
+            
+            destination = self.bot.get_channel(session.channel)
+            
+            msg = await destination.fetch_message(session.messages[message.id])
+            await msg.delete()
 
     @commands.Cog.listener("on_typing")
     async def on_typing(self, channel, user, when):
@@ -95,8 +171,13 @@ class Linker(commands.Cog):
         if not session:
             return
         
-        #Trigger typing because someone is typing in the linked channel
-        await self.bot.get_channel(session.channel).trigger_typing()
+        if isinstance(session, Session):
+            #Trigger typing because someone is typing in the linked channel
+            await self.bot.get_channel(session.channel).trigger_typing()
+
+        if isinstance(session, DMSession):
+            #Triger typing because the owner is typing
+            await self.bot.get_channel(session.channel).trigger_typing()
 
     @commands.Cog.listener("on_message_edit")
     async def on_message_edit(self, before, after):
@@ -107,21 +188,29 @@ class Linker(commands.Cog):
         if not session:
             return
         
-        #Since webhooks can't edit messages, return if the message isn't on the bottom
-        if before.channel.last_message_id != before.id:
-            return False
+        if isinstance(session, Session):
+            #Since webhooks can't edit messages, return if the message isn't on the bottom
+            if before.channel.last_message_id != before.id:
+                return False
 
-        destination = self.bot.get_channel(session.channel)
+            destination = self.bot.get_channel(session.channel)
 
-        #Get the message to edit, which we delete
-        #Then send the after message with "(edited)" on the end
-        to_delete = await destination.fetch_message(session.messages[before.id])
+            #Get the message to edit, which we delete
+            #Then send the after message with "(edited)" on the end
+            to_delete = await destination.fetch_message(session.messages[before.id])
 
-        await to_delete.delete()
-        
-        #Resend the message and add it to the message dict
-        msg = await session.webhook.send(f"{after.content}  ⁽ᵉᵈᶦᵗᵉᵈ⁾", username=before.author.display_name, avatar_url=before.author.avatar_url, wait=True)
-        session.messages[before.id] = msg.id
+            await to_delete.delete()
+            
+            #Resend the message and add it to the message dict
+            msg = await session.webhook.send(f"{after.content}  ⁽ᵉᵈᶦᵗᵉᵈ⁾", username=before.author.display_name, avatar_url=before.author.avatar_url, wait=True)
+            session.messages[before.id] = msg.id
+
+        if isinstance(session, DMSession):
+            destination = self.bot.get_channel(session.channel)
+
+            #Get the message and edit it
+            msg = await destination.fetch_message(session.messages[before.id])
+            await msg.edit(content=after.content)
 
 def setup(bot):
     bot.add_cog(Linker(bot))
